backend/requirements.txt
langchain==0.2.17
langgraph
arxiv
rank-bm25
numpy
scikit-learn
pydantic
fastapi
uvicorn
python-dotenv
requests
beautifulsoup4
sentence-transformers
yandexcloud
websockets
